Use create instead of createOrUpdate (#43)
We were using createOrUpdate to make sure we overwrite any stray lease. But this was unnecessary extra caution as we anyway explicitly delete any stray lease. The problem with createOrUpdate is that for the create case, we need to provide a some version to avoid overwriting a legit lease, which isn't officially supported.

Co-authored-by: Jonas Chapuis <jonas.chapuis@zf.com>